- 调整代码

+++ Engine/Resource/src/ResourceLoaderTexture.cpp
@@ -18,14 +18,14 @@ static bool InternalLoadTexture2D(CGfxTexture2D* pTexture2D, const gli::texture*
 	}
 
 	if (gli::is_compressed(texture->format())) {
-		for (int level = baseLevel; level < (int)texture->levels(); level++) {
+		for (int level = baseLevel; level < texture->levels(); level++) {
 			if (pTexture2D->TransferTexture2DCompressed((GfxPixelFormat)texture->format(), level - baseLevel, 0, 0, texture->extent(level).x, texture->extent(level).y, texture->size(level), texture->data(0, 0, level)) == false) {
 				return false;
 			}
 		}
 	}
 	else {
-		for (int level = baseLevel; level < (int)texture->levels(); level++) {
+		for (int level = baseLevel; level < texture->levels(); level++) {
 			if (pTexture2D->TransferTexture2D((GfxPixelFormat)texture->format(), level - baseLevel, 0, 0, texture->extent(level).x, texture->extent(level).y, GFX_DATA_UNSIGNED_BYTE, texture->size(level), texture->data(0, 0, level)) == false) {
 				return false;
 			}
@@ -49,8 +49,8 @@ static bool InternalLoadTexture2DArray(CGfxTexture2DArray* pTexture2DArray, cons
 	}
 
 	if (gli::is_compressed(texture->format())) {
-		for (int layer = 0; layer < (int)texture->layers(); layer++) {
-			for (int level = baseLevel; level < (int)texture->levels(); level++) {
+		for (int layer = 0; layer < texture->layers(); layer++) {
+			for (int level = baseLevel; level < texture->levels(); level++) {
 				if (pTexture2DArray->TransferTexture2DCompressed((GfxPixelFormat)texture->format(), layer, level - baseLevel, 0, 0, texture->extent(level).x, texture->extent(level).y, texture->size(level), texture->data(layer, 0, level)) == false) {
 					return false;
 				}
@@ -58,8 +58,8 @@ static bool InternalLoadTexture2DArray(CGfxTexture2DArray* pTexture2DArray, cons
 		}
 	}
 	else {
-		for (int layer = 0; layer < (int)texture->layers(); layer++) {
-			for (int level = baseLevel; level < (int)texture->levels(); level++) {
+		for (int layer = 0; layer < texture->layers(); layer++) {
+			for (int level = baseLevel; level < texture->levels(); level++) {
 				if (pTexture2DArray->TransferTexture2D((GfxPixelFormat)texture->format(), layer, level - baseLevel, 0, 0, texture->extent(level).x, texture->extent(level).y, GFX_DATA_UNSIGNED_BYTE, texture->size(level), texture->data(layer, 0, level)) == false) {
 					return false;
 				}
@@ -80,14 +80,14 @@ static bool InternalLoadTexture2DArrayLayer(CGfxTexture2DArray* pTexture2DArray,
 	}
 
 	if (gli::is_compressed(texture->format())) {
-		for (int level = baseLevel; level < (int)texture->levels(); level++) {
+		for (int level = baseLevel; level < texture->levels(); level++) {
 			if (pTexture2DArray->TransferTexture2DCompressed((GfxPixelFormat)texture->format(), layer, level - baseLevel, 0, 0, texture->extent(level).x, texture->extent(level).y, texture->size(level), texture->data(0, 0, level)) == false) {
 				return false;
 			}
 		}
 	}
 	else {
-		for (int level = baseLevel; level < (int)texture->levels(); level++) {
+		for (int level = baseLevel; level < texture->levels(); level++) {
 			if (pTexture2DArray->TransferTexture2D((GfxPixelFormat)texture->format(), layer, level - baseLevel, 0, 0, texture->extent(level).x, texture->extent(level).y, GFX_DATA_UNSIGNED_BYTE, texture->size(level), texture->data(0, 0, level)) == false) {
 				return false;
 			}
@@ -111,7 +111,7 @@ static bool InternalLoadTextureCubeMap(CGfxTextureCubeMap* pTextureCubeMap, cons
 	}
 
 	if (gli::is_compressed(texture->format())) {
-		for (int level = baseLevel; level < (int)texture->levels(); level++) {
+		for (int level = baseLevel; level < texture->levels(); level++) {
 			if (pTextureCubeMap->TransferTexture2DCompressed((GfxPixelFormat)texture->format(), GFX_TEXTURE_CUBEMAP_POSITIVE_X, level - baseLevel, 0, 0, texture->extent(level).x, texture->extent(level).y, texture->size(level), texture->data(0, 0, level)) == false) {
 				return false;
 			}
@@ -133,7 +133,7 @@ static bool InternalLoadTextureCubeMap(CGfxTextureCubeMap* pTextureCubeMap, cons
 		}
 	}
 	else {
-		for (int level = baseLevel; level < (int)texture->levels(); level++) {
+		for (int level = baseLevel; level < texture->levels(); level++) {
 			if (pTextureCubeMap->TransferTexture2D((GfxPixelFormat)texture->format(), GFX_TEXTURE_CUBEMAP_POSITIVE_X, level - baseLevel, 0, 0, texture->extent(level).x, texture->extent(level).y, GFX_DATA_UNSIGNED_BYTE, texture->size(level), texture->data(0, 0, level)) == false) {
 				return false;
 			}
@@ -168,14 +168,14 @@ static bool InternalLoadTextureCubeMapFace(CGfxTextureCubeMap* pTextureCubeMap, 
 	}
 
 	if (gli::is_compressed(texture->format())) {
-		for (int level = baseLevel; level < (int)texture->levels(); level++) {
+		for (int level = baseLevel; level < texture->levels(); level++) {
 			if (pTextureCubeMap->TransferTexture2DCompressed((GfxPixelFormat)texture->format(), face, level - baseLevel, 0, 0, texture->extent(level).x, texture->extent(level).y, texture->size(level), texture->data(0, 0, level)) == false) {
 				return false;
 			}
 		}
 	}
 	else {
-		for (int level = baseLevel; level < (int)texture->levels(); level++) {
+		for (int level = baseLevel; level < texture->levels(); level++) {
 			if (pTextureCubeMap->TransferTexture2D((GfxPixelFormat)texture->format(), face, level - baseLevel, 0, 0, texture->extent(level).x, texture->extent(level).y, GFX_DATA_UNSIGNED_BYTE, texture->size(level), texture->data(0, 0, level)) == false) {
 				return false;
 			}
